Fixed the delegation of the new people adder view, kinda
Check out the comments in the showContacts method for details.
Basically everything works now, but I got a bit creative with the
delegation. Unless I have no idea what I'm doing and that's how you're
supposed to do it anyway.

I also removed a bunch of the Three20 snippets left over and
reorganized a bit.